Move ThreadsafeRestrictionTracker and RestrictionTrackerView out from iobase.py
